% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdistpara.R
\name{hdistpara}
\alias{hdistpara}
\title{Parametric estimation of the Hellinger distance between two random variates}
\usage{
hdistpara(x1, x2 = NULL, params = NULL, densfun = "normal")
}
\arguments{
\item{x1}{A \code{numeric} random variate of draws from the first
distribution}

\item{x2}{An optional \code{numeric} random variate of draws from the second
distribution.}

\item{params}{An optional \code{numeric} of two parameters giving the
parameters for the density function.}

\item{minx}{A \code{character} giving the density function to fit to both
variates. Currently only "normal" and "beta" are supported.}
}
\value{
A helldistp object containing approximate Hellinger distances and
 fitted density kernals.
\item{hdist}{Estimate of Hellinger distance}
}
\description{
Parametric estimation of the Hellinger distance between two random variates
}
\details{
Hellinger distance is approximated by fitting distributions using
 \code{MASS::fitdistr} and then calculating the exact Hellinger distance
given the fitted parameters. Currently the only options are to compare two Beta distributions or two normal distributions (the default).

If \code{params} is given the second density function will be specified
exactly. If \code{x2} is given, the second density function will be estimate from the random variate.
If using \code{params} the parameters should be the mean and sd
(ie \code{c(mean, sd)}) in that order for 'normal' density and a and
b for the beta distribution (ie \code{c(a, b)}) in that order.
'
Class \item{helldistp} has a plot method that can be used to compared
the discrete and continuous distribution fits. It is recommended to
visually check distribution fits, particularly if the number of random
variates is small.

In general these methods will be inaccurate if analysis is performed on too
few samples, e.g. <10 000. >100 000 would be ideal.
}
\author{
Christopher J. Brown christo.j.brown@gmail.com
}
