% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellinger.R
\name{hellinger}
\alias{hellinger}
\title{Estimate the Hellinger distance between two random variates}
\usage{
hellinger(x1, x2, nbreaks = 100, minx = min(c(x1, x2)), maxx = max(c(x1,
  x2)))
}
\arguments{
\item{x1}{A \code{numeric} random variate of draws from the first distribution}

\item{x2}{A \code{numeric} random variate of draws from the second distribution}

\item{nbreaks}{A single \code{numeric} giving how many breaks to break the
discrete distribution into}

\item{minx}{A single \code{numeric} giving the lower bound of integration}

\item{maxx}{A single \code{numeric} giving the upper bound of integration}
}
\value{
A helldist object containing approximate Hellinger distances and
fitted density kernals.
\item{hdist_disc}{Estimate of Hellinger distance using discrete approximation
of the distributions}
\item{hdist_cont}{Estimate of Hellinger distance using continous approximation
of distributions}
}
\description{
Estimate the Hellinger distance between two random variates
}
\details{
Hellinger distance is approximated in two ways:

(1) by binning the random variates and calculating the Hellinger distance
for discrete distributions and

(2) by creating a continuous approximation of the distributions using
\code{density} and then using numerical integration to calculate the
Hellinger distance.

Method (2) - continuous integration - should in genernal be more accurate
however, it may give poor approximations for multi-modal distributions.

Continuous integration may return NaN if the distributions are near identical.
Class \item{helldist} has a plot method that can be used to compared the
discrete and continuous distribution fits.

It is recommended to visually
check distribution fits, particularly if the number of random variates is
small.

In general these methods will be inaccurate if analysis is performed on
too few samples, e.g. <10 000. >100 000 would be ideal.
}
\author{
Christopher J. Brown christo.j.brown@gmail.com
}

