% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeSens.R
\docType{package}
\name{BayeSens-package}
\alias{BayeSens}
\alias{BayeSens-package}
\title{Tools for prior sensitivity analysis with Bayesian MCMC models}
\description{
The \pkg{HellingerMCMC} provides functions for non-parametric calculation of the Hellinger distance, \code{\link{hellinger}}, parametric calculation of the Hellinger distnace \code{\link{hdistpara}}; data cloning with \code{\link{dataclone}}; and calculating posterior shrinkage with \code{\link{postshrink}}. Refer to the vignette for more help by typing \code{vignette('BayeSens')}.
}

